e2e: extracting fencing credentials rotaion into standalone test
